Add support to Elide for coercing text values to Java 8 time classes.
This enhancement is required because the Apache Commons BeanUtils library used by the Elide CoerceUtil class currently does not support the Java 8 time classes.

This enhancement uses the parse() method provided with each of the Java 8 time classes to perform the conversion. A better solution might be to use Jackson deserialization, but that does not appear to be a straightforward change with the way CoerceUtil is currently structured.